processors/SpeechProcessor/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
attrs==25.4.0
audioread==3.1.0
babel==2.17.0
bangla==0.0.5
blinker==1.9.0
blis==1.3.2
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.13
Cython==3.2.1
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
filelock==3.20.0
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
g2pkk==0.1.2
grpcio==1.76.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.36.0
idna==3.11
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.15
networkx==2.8.8
nltk==3.9.2
num2words==0.5.14
numba==0.62.1
numpy==1.26.4
openai-whisper==20250625
packaging==25.0
pandas==1.5.3
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
preshed==3.0.11
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pynndescent==0.5.13
pyparsing==3.2.5
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
smart_open==7.5.0
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.9
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20251022
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.9
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
tqdm==4.67.1
trainer==0.0.36
transformers==4.57.1
TTS==0.22.0
typeguard==4.4.4
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
Unidecode==1.4.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.3
wrapt==2.0.1
yarl==1.22.0
